Modernization and Code Quality #2
Changes:
- Adjusted README to new updated version of a2lparser
- Changes pyproject.toml for pypi
- Improved CLI with code library (removed direct eval call)
- Moved converts from a2l subfolder to converters under root package folder
