+++ .github/workflows/build.yml
@@ -35,15 +35,19 @@ jobs:
       # Install build dependencies
       - name: Install build dependencies
         run: |
-          python -m pip install setuptools wheel
+          python -m pip install setuptools wheel build
       # Building the package
       - name: Build package
         run: |
-          python -m pip install .
+          python -m build
+      # Installing package
+      - name: Install package
+        run: |
+          pip install dist\*.whl
       # Running a2lparser.
       - name: Print package version
         run: |
-          a2lparser --version
+          a2lparser --version || exit /b 1
 
 
   ###############
@@ -59,7 +63,6 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: 3.11
-
       # Set up virtual environment
       - name: Set up virtual environment
         run: |
@@ -67,18 +70,19 @@ jobs:
           python -m venv .venv
       - name: Activate virtual environment
         run: source .venv/bin/activate
-
       # Install build dependencies
       - name: Install build dependencies
         run: |
-          python -m pip install setuptools wheel
-
+          python -m pip install setuptools wheel build
       # Building the package
       - name: Build package
         run: |
-          python -m pip install .
-
+          python -m build
+      # Installing package
+      - name: Install package
+        run: |
+          pip install dist/*.whl
       # Running a2lparser.
       - name: Print package version
         run: |
-          a2lparser --version
+          a2lparser --version || exit 1
