+++ a2lparser/converter/yaml_converter.py
@@ -20,7 +20,7 @@
 
 
 import yaml
-from a2lparser.a2l.converter.a2l_converter import A2LConverter
+from a2lparser.converter.a2l_converter import A2LConverter
 
 
 class YAMLConverter(A2LConverter):
